feat: add replyToMessageId to CreateMessageFields
Add a replyToMessageId field to CreateMessageFields, enabling users
to reply to a specific message when creating a plain (that is,
non-interaction) message.

Please note that this patch is entirely untested in terms of if it works
when used in scripts, I only tried compiling it, which it does.

That said, I presume the API will look like this in the
`CreateMessageFields` interface:

```typescript
// ...
    replyToMessageId?: "messageID",
// ...
```

Thus, handing such an interface to Discord.createMessage() should result
in a message replying to some other message in that channel with that
ID.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>