feat: massive restructuring and cleanup of scheduler and vm_worker
Most of this was written once years ago to make it work,  then regurlarly tweaked around as new bugs came up.

The problem was that i never wanted to fix the bad types or the bad flow, the shutdown event for example was unreliable and decoupled from the actual closing of the event channel, the shutdown types themselves were weird and so on.

This also adds the tracking of the current vm session index so that we only throw away the pending acks of the sessions that were shut down, this removes some jank workaround for the previous logic

In general it's much improved, but it needs more testing since this touches on the very core parts of the bot.